Bug fixes, config file, average
Fixed all times (in ms) to use long, rather than int.  (16-bit
overflows).

Added code to read min time between samples from /config.txt

Added "average of all the data packets since last time" to the file --
useful when you set a min time between samples.

+++ button.h
@@ -9,13 +9,13 @@ class button
     void clear_new_push(void);
 
   private:
-    int m_pin;
-    int m_last_down_ms;
-    int m_last_up_ms;
+    int     m_pin;
+    long    m_last_down_ms;
+    long    m_last_up_ms;
     boolean m_wait_for_up;
     boolean m_pushed;
     boolean m_new_push;
-    int m_new_push_at_ms;
+    long    m_new_push_at_ms;
 };
